% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adverbial.R
\name{adverbial}
\alias{adverbial}
\alias{nom_adv}
\alias{nom_times}
\title{Convert numbers to adverbial character vectors (once, twice, three times)}
\usage{
adverbial(x, thrice = FALSE, cardinal = TRUE, max_n = Inf, ...)

nom_adv(x, thrice = FALSE, cardinal = TRUE, max_n = Inf, ...)

nom_times(x, thrice = FALSE, cardinal = TRUE, max_n = Inf, ...)
}
\arguments{
\item{x}{A numeric vector}

\item{thrice}{A logical of length one.
If \code{TRUE}, the adverbial of \code{3} will be "thrice".
If \code{FALSE}, the adverbial of \code{3} will be "three times".
Defaults to \code{FALSE}.}

\item{cardinal}{Whether to convert a numeric vector with \code{\link[=cardinal]{cardinal()}}
before applying adverbial suffixes.
When \code{TRUE}, 1 -> "once".
When \code{FALSE}, 1 -> "1 time".
Defaults to \code{TRUE}.}

\item{max_n}{A numeric vector.
When the absolute value of \code{x} is greater than \code{max_n}, \code{x} remains
numeric instead of being converted to words.
If \code{max_n} is negative, no \code{x}s will be converted to words.
Defaults to \code{Inf}, which converts all \code{x}s to words.}

\item{...}{Additional arguments passed to \code{\link[=cardinal]{cardinal()}}}
}
\value{
A character vector of the same length as \code{x}
}
\description{
Convert numbers to adverbial character vectors (once, twice, three times)
}
\examples{
nom_adv(1:4)
nom_adv(1:4, thrice = TRUE)
}
\seealso{
Other number names: 
\code{\link{cardinal}()},
\code{\link{collective}()},
\code{\link{denominator}()},
\code{\link{numerator}()},
\code{\link{ordinal}()},
\code{\link{ratio}()}
}
\concept{number names}
