% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardinal.R
\name{cardinal}
\alias{cardinal}
\title{Convert numbers to cardinal character vectors}
\usage{
cardinal(x, negative = getOption("numerate.negative", "negative"))
}
\arguments{
\item{x}{A numeric vector}

\item{negative}{A character to append to negative numbers.
Defaults to \code{"negative"}.
Default can be changed by setting \code{options("numerate.negative")}.}
}
\value{
A character vector of the same length as \code{x}
}
\description{
Convert numbers to cardinal character vectors
}
\examples{
cardinal(2)
cardinal(1:10)
cardinal(2 + 4/9)
cardinal(-2)
cardinal(-2, negative = "minus")

cardinal(525600)
paste("There are", cardinal(525600), "minutes in a year.")

cardinal(3.72e13)
paste("There are", cardinal(3.72e13), "cells in the human body.")
}
