% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collective.R
\name{collective}
\alias{collective}
\alias{nom_coll}
\title{Convert numbers to collective character vectors (the, both, all three)}
\usage{
collective(x, ...)

nom_coll(x, ...)
}
\arguments{
\item{x}{A numeric vector}

\item{...}{Additional arguments passed to \code{\link[=cardinal]{cardinal()}}}
}
\value{
A character vector of the same length as \code{x}
}
\description{
Convert numbers to collective character vectors (the, both, all three)
}
\examples{
paste(nom_coll(0:3), "fish")
paste(nom_coll(9:12, max_n = 10), "fish")
}
\seealso{
Other number names: 
\code{\link{adverbial}()},
\code{\link{cardinal}()},
\code{\link{denominator}()},
\code{\link{ordinal}()}
}
\concept{number names}
