% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denominator.R
\name{denominator}
\alias{denominator}
\title{Convert numbers to ordinal character vectors}
\usage{
denominator(
  x,
  numerator = 1,
  quarter = getOption("numerate.quarter", TRUE),
  negative = getOption("numerate.negative", "negative")
)
}
\arguments{
\item{x}{A numeric vector}

\item{numerator}{A numeric vector.
The numerator(s) associated with the denominator(s).
When \code{numerator} is \code{1} or \code{-1}, returns a singular denominator.
When \code{numerator} is not \code{1} or \code{-1}, returns a plural denominator.}

\item{quarter}{A logical of length one.
If \code{TRUE}, the denominator of \code{4} will be "quarter(s)".
If \code{FALSE}, the denominator of \code{4} will be "fourth(s)".
Defaults to \code{TRUE}.
Default can be changed by setting \code{options("numerate.quarter")}.}
}
\value{
A character vector of the same length as \code{x}
}
\description{
Convert numbers to ordinal character vectors
}
\examples{
denominator(2)
denominator(1:10)
denominator(1:10, numerator = 2)
denominator(1:10, numerator = 1:10)

denominator(4)
denominator(4, quarter = FALSE)
}
