% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nombre-package.R
\docType{package}
\name{nombre-package}
\alias{nombre}
\alias{nombre-package}
\title{nombre: Number Names}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Converts numeric vectors to character vectors of
    English number names. Provides conversion to cardinals, ordinals,
    numerators, and denominators. Supports negative and non-integer
    numbers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nombre.rossellhayes.com}
  \item \url{https://github.com/rossellhayes/nombre}
  \item Report bugs at \url{https://github.com/rossellhayes/nombre/issues}
}

}
\author{
\strong{Maintainer}: Alexander Rossell Hayes \email{alexander@rossellhayes.com} (\href{https://orcid.org/0000-0001-9412-0457}{ORCID})

}
\keyword{internal}
