% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerator.R
\name{numerator}
\alias{numerator}
\alias{nom_numer}
\title{Convert numbers to numerator character vectors (one, two, three)}
\usage{
numerator(x, ...)

nom_numer(x, ...)
}
\arguments{
\item{x}{A numeric vector}

\item{...}{Additional arguments passed to \code{\link[=cardinal]{cardinal()}}}
}
\description{
\code{nom_numer()} and \code{numerator()} are equivalent to \code{\link[=nom_card]{nom_card()}} and
\code{\link[=cardinal]{cardinal()}} for integers, but \link{cardinal}s support fractional components
while \code{numerator}s do not.
}
\seealso{
Other number names: 
\code{\link{adverbial}()},
\code{\link{cardinal}()},
\code{\link{collective}()},
\code{\link{denominator}()},
\code{\link{ordinal}()},
\code{\link{ratio}()}
}
\concept{number names}
