% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio.R
\name{ratio}
\alias{ratio}
\alias{nom_ratio}
\title{Convert numbers to ratio character vectors (two to one, one in three, five out of ten)}
\usage{
ratio(
  x,
  sep = get_config("nombre::sep", "to"),
  max_n = get_config("nombre::max_n", Inf),
  negative = get_config("nombre::negative", "negative"),
  common_denom = get_config("nombre::common_denom", FALSE),
  ...
)

nom_ratio(
  x,
  sep = get_config("nombre::sep", "to"),
  max_n = get_config("nombre::max_n", Inf),
  negative = get_config("nombre::negative", "negative"),
  common_denom = get_config("nombre::common_denom", FALSE),
  ...
)
}
\arguments{
\item{x}{A numeric vector}

\item{sep}{A character vector separating components of the ratio. \code{sep} must
be "in", "of", "to", or "out of".
Defaults to \code{"to"}.
Default can be changed with \link[=set_config]{set_config("nombre::sep")}.}

\item{max_n}{A numeric vector.
When the absolute value of \code{x} is greater than \code{max_n}, \code{x} remains
numeric instead of being converted to words.
If \code{max_n} is negative, no \code{x}s will be converted to words.
(This can be useful when \code{max_n} is passed by another function.)
Defaults to \code{Inf}, which converts all \code{x}s to words.
Default can be changed with \link[=set_config]{set_config("nombre::max_n")}.}

\item{negative}{A character vector to append to negative numbers.
Defaults to \code{"negative"}.
Default can be changed with
\link[=set_config]{set_config("nombre::negative")}.}

\item{common_denom}{Logical. If TRUE, all ratios use a common denominator
value. If FALSE, a denominator of 1 is used when \code{x} is 0. Default can be
changed with \link[=set_config]{set_config("nombre::common_denom")}.}

\item{...}{Additional arguments passed to \code{\link[fracture:frac_mat]{fracture::frac_mat()}}.
See details.}
}
\value{
A character vector of the same length as \code{x}
}
\description{
Convert numbers to ratio character vectors (two to one, one in three, five out of ten)
}
\details{
\code{x} is converted to a fraction by \code{\link[fracture:frac_mat]{fracture::frac_mat()}}.
Named arguments of \code{ratio()} and \code{nom_ratio()} are passed to
\code{\link[fracture:frac_mat]{fracture::frac_mat()}} through \code{...}.
Helpful arguments include:
\itemize{
\item \code{base_10 = TRUE}, which forces all fractions to use denominators that are
powers of ten.
\item \code{max_denom}, which sets the maximum allowable denominator.
By default, the maximum denominator is "ten-millionths".
}

The \code{\link[fracture:frac_mat]{fracture::frac_mat()}}  argument \code{mixed} is set to TRUE.
}
\examples{
paste0("Our team is outnumbered ", nom_ratio(10), ".")
paste0("The chances of winning are ", nom_ratio(1/1000000, sep = "in"), ".")

nom_ratio(c(1, 10, 100))
nom_ratio(c(0, 0.5, 1.5))
nom_ratio(c(0, 0.125, 0.625, 1), sep = "out of", common_denom = TRUE)
nom_ratio(5 / 10, sep = "in", base_10 = TRUE)
nom_ratio(6 / 25, sep = "in")
nom_ratio(6 / 25, sep = "out of", max_denom = 10)
}
\seealso{
Other number names: 
\code{\link{adverbial}()},
\code{\link{cardinal}()},
\code{\link{collective}()},
\code{\link{denominator}()},
\code{\link{numerator}()},
\code{\link{ordinal}()}
}
\concept{number names}
