% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncardinal.R
\name{uncardinal}
\alias{uncardinal}
\alias{nom_uncard}
\title{Convert cardinal character vectors to numbers}
\usage{
uncardinal(x)

nom_uncard(x)
}
\arguments{
\item{x}{A character vector of the cardinal names of numbers}
}
\value{
A numeric vector the same length as \code{n}.
NAs will be produced for numbers with fractions or decimals or non-cardinal
numbers (e.g. ordinals).
}
\description{
This function is in experimental development.
It currently only supports English cardinal integers or character vectors
produced by one of \link{nombre}'s functions.
}
\examples{
uncardinal("one")
uncardinal("negative one hundred fifty-seven")
uncardinal(
  c(
    "twenty-five",
    "one million two hundred thirty-four thousand five hundred sixty-seven"
  )
)
uncardinal("infinity")

card <- cardinal(25)
uncardinal(card)
ord <- ordinal(25)
uncardinal(ord)
}
\seealso{
\code{\link[=cardinal]{cardinal()}} to convert numeric vectors to number names
}
