% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardinal.R
\name{cardinal}
\alias{cardinal}
\alias{nmr_card}
\alias{numerator}
\alias{nmr_num}
\title{Convert numbers to cardinal character vectors}
\usage{
cardinal(x, negative = getOption("numerate.negative", "negative"))

nmr_card(x, negative = getOption("numerate.negative", "negative"))

numerator(x, negative = getOption("numerate.negative", "negative"))

nmr_num(x, negative = getOption("numerate.negative", "negative"))
}
\arguments{
\item{x}{A numeric vector}

\item{negative}{A character to append to negative numbers.
Defaults to \code{"negative"}.
Default can be changed by setting \code{options("numerate.negative")}.}
}
\value{
A character vector of the same length as \code{x}
}
\description{
\code{cardinal()} and \code{nmr_card()} produce cardinal numbers.
\code{numerator()} and \code{nmr_num()} produce numerators.
The results are equivalent for integers, but \code{cardinal()} and \code{nmr_card()}
support fractional components while \code{numerator()} and \code{nmr_num()} do not.
}
\examples{
nmr_card(2)
nmr_card(1:10)
nmr_card(2 + 4/9)
nmr_card(-2)
nmr_card(-2, negative = "minus")

nmr_card(525600)
paste("There are", nmr_card(525600), "minutes in a year.")

nmr_card(3.72e13)
paste("There are", nmr_card(3.72e13), "cells in the human body.")
}
