% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collective.R
\name{collective}
\alias{collective}
\alias{nom_coll}
\title{Convert numbers to collective character vectors (the, both, all three)}
\usage{
collective(
  x,
  all_n = get_config("nombre::all_n", TRUE),
  of_the = get_config("nombre::of_the", FALSE),
  ...
)

nom_coll(
  x,
  all_n = get_config("nombre::all_n", TRUE),
  of_the = get_config("nombre::of_the", FALSE),
  ...
)
}
\arguments{
\item{x}{A numeric vector.}

\item{of_the}{Whether to include "of the" for collectives other than 1.
Defaults to \code{FALSE}.
Default can be changed with \link[=set_config]{set_config("nombre::of")}.}

\item{...}{Additional arguments passed to \code{\link[=cardinal]{cardinal()}}.}
}
\value{
A character vector of the same length as \code{x}.
}
\description{
Convert numbers to collective character vectors (the, both, all three)
}
\examples{
paste(nom_coll(0:3), "fish")
paste(nom_coll(9:12, max_n = 10), "fish")
}
\seealso{
Other number names: 
\code{\link{adverbial}()},
\code{\link{cardinal}()},
\code{\link{denominator}()},
\code{\link{numerator}()},
\code{\link{ordinal}()}
}
\concept{number names}
