% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal.R
\name{ordinal}
\alias{ordinal}
\title{Convert numbers to ordinal character vectors}
\usage{
ordinal(x, negative = getOption("numerate.negative", "negative"))
}
\arguments{
\item{x}{A numeric vector}

\item{negative}{A character of length one to append to negative numbers.
Defaults to \code{"negative"}.
Default can be changed by setting \code{options("numerate.negative")}.}
}
\value{
A character vector of the same length as \code{x}
}
\description{
Convert numbers to ordinal character vectors
}
